Moving Color unit tests next to the others
