--- Tests/ColorTests/nano.runsettings
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RunSettings>
-   <!-- Configurations that affect the Test Framework -->
-   <RunConfiguration>
-       <ResultsDirectory>.\TestResults</ResultsDirectory><!-- Path relative to solution directory -->
-       <TestSessionTimeout>120000</TestSessionTimeout><!-- Milliseconds -->
-       <TargetFrameworkVersion>net48</TargetFrameworkVersion>
-       <TargetPlatform>x64</TargetPlatform>
-   </RunConfiguration>
-   <nanoFrameworkAdapter>
-       <Logging>None</Logging> <!--Set to the desired level of logging for Unit Test execution. Possible values are: None, Detailed, Verbose, Error. -->
-       <IsRealHardware>False</IsRealHardware><!--Set to true to run tests on real hardware. -->
-   </nanoFrameworkAdapter>
-</RunSettings>
