+++ nanoFramework.Graphics.sln
@@ -45,8 +45,6 @@ Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.Graphics.Core
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{DEAF4062-7DF3-4DDC-B4A0-F37C87865951}"
 EndProject
-Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "ColorTests", "Tests\ColorTests\ColorTests.nfproj", "{CC585116-52D9-4FA0-BF8D-6F8E1C9894FF}"
-EndProject
 Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.Graphics.Core.UnitTests", "nanoFramework.Graphics.Core.UnitTests\nanoFramework.Graphics.Core.UnitTests.nfproj", "{7B17788C-3B57-4EA4-A391-4FB7BBE0BE07}"
 EndProject
 Global
@@ -115,12 +113,6 @@ Global
 		{67CD2B6C-E71A-4A98-9C7C-B88109ABE158}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{67CD2B6C-E71A-4A98-9C7C-B88109ABE158}.Release|Any CPU.Build.0 = Release|Any CPU
 		{67CD2B6C-E71A-4A98-9C7C-B88109ABE158}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{CC585116-52D9-4FA0-BF8D-6F8E1C9894FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CC585116-52D9-4FA0-BF8D-6F8E1C9894FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CC585116-52D9-4FA0-BF8D-6F8E1C9894FF}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{CC585116-52D9-4FA0-BF8D-6F8E1C9894FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CC585116-52D9-4FA0-BF8D-6F8E1C9894FF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CC585116-52D9-4FA0-BF8D-6F8E1C9894FF}.Release|Any CPU.Deploy.0 = Release|Any CPU
 		{7B17788C-3B57-4EA4-A391-4FB7BBE0BE07}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7B17788C-3B57-4EA4-A391-4FB7BBE0BE07}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7B17788C-3B57-4EA4-A391-4FB7BBE0BE07}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
@@ -140,7 +132,6 @@ Global
 		{61918B13-7E23-4567-8A4B-FE1F0248D0ED} = {B56DBFF9-1C54-4D42-8DC7-B6E35FE56268}
 		{EDE1D4FC-4F60-41E2-B3CC-A0671662A3C8} = {B56DBFF9-1C54-4D42-8DC7-B6E35FE56268}
 		{6DC496E5-22D2-4138-A5A0-6FCE8BF3785D} = {B56DBFF9-1C54-4D42-8DC7-B6E35FE56268}
-		{CC585116-52D9-4FA0-BF8D-6F8E1C9894FF} = {DEAF4062-7DF3-4DDC-B4A0-F37C87865951}
 		{7B17788C-3B57-4EA4-A391-4FB7BBE0BE07} = {DEAF4062-7DF3-4DDC-B4A0-F37C87865951}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
